# apps/api/openapi/spec.yaml
openapi: 3.0.0
info:
  title: lomi.africa Payments API
  version: 1.0.0
  description: API of lomi. payment orchestration platform
servers:
  - url: https://api.lomi.africa/v1
    
# API Schema to Database Table Mapping:
# - Merchant (API) -> merchants (DB)
# - Product (API) -> merchant_products (DB)
# - SubscriptionPlan (API) -> subscription_plans (DB)
# - Subscription (API) -> merchant_subscriptions (DB)
# - Transaction (API) -> transactions (DB)
# - Customer (API) -> customers (DB)
# - PaymentLink (API) -> payment_links (DB)
# - Refund (API) -> refunds (DB)
# - Webhook (API) -> webhooks (DB)
# - CheckoutSession (API) -> checkout_sessions (DB)
# - MerchantAccount (API) -> merchant_accounts (DB)
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  responses:
    UnauthorizedError:
      description: API key is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Merchant:
      type: object
      properties:
        merchant_id:
          type: string
          format: uuid
        name: 
          type: string
        email:
          type: string
          format: email
        phone_number:
          type: string
        organization_id:
          type: string
          format: uuid
        country:
          type: string
        mrr:
          type: number
        arr:
          type: number
        merchant_lifetime_value:
          type: number
        retry_payment_every:
          type: integer
        total_retries:
          type: integer
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PaymentLinkType:
      type: string
      enum:
        - product
        - plan
        - instant
...
security:
  - ApiKeyAuth: []
paths:
...
  /merchants/{merchant_id}:
    get:
      summary: Get merchant details
      parameters:
        - name: merchant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Merchant'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /merchants/{merchant_id}/providers:
    get:
      summary: List connected payment providers for a merchant
      operationId: merchantProviders
      parameters:
        - name: merchant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConnectedProvider'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /products:
    post:
      summary: Create a new product
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProduct'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
                
    get:
      summary: List products for a merchant
      parameters:
        - name: merchant_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /products/{product_id}:
    get:
      summary: Get product details
      description: Get details of a specific product
      operationId: getProduct
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      summary: Update product status
      description: Update only the active status of a product. Core details (name, price, etc.) cannot be changed after creation.
      operationId: updateProductStatus
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                is_active:
                  type: boolean
              required:
                - is_active
      responses:
        '200':
          description: Product status updated successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Delete product
      description: Delete a product
      operationId: deleteProduct
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Product deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /merchants/{merchant_id}/subscriptions:
    post:
      summary: Create a subscription plan for a merchant
      parameters:
        - name: merchant_id
          in: path
          required: true
          description: The ID of the merchant for whom to create the plan.
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionPlan'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionPlan'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
                
    get:
      summary: List subscription plans for a merchant  
      parameters:
        - name: merchant_id
          in: path
          required: true
          description: The ID of the merchant whose plans to list.
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
        - name: offset
          in: query
          required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionPlan'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /merchants/{merchant_id}/subscriptions/{plan_id}:
    get:
      summary: Get subscription plan details for a merchant
      description: Get details of a specific subscription plan belonging to a merchant
      operationId: getSubscriptionPlan
      parameters:
        - name: merchant_id
          in: path
          required: true
          description: The ID of the merchant.
          schema:
            type: string
            format: uuid
        - name: plan_id
          in: path
          required: true
          description: The ID of the subscription plan.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Subscription plan details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionPlan'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      summary: Update subscription plan
      description: Update specific fields (is_active, metadata) of a subscription plan. Core details are immutable via API.
      operationId: updateSubscriptionPlan
      parameters:
        - name: merchant_id
          in: path
          required: true
          description: The ID of the merchant.
          schema:
            type: string
            format: uuid
        - name: plan_id
          in: path
          required: true
          description: The ID of the subscription plan.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                is_active:
                  type: boolean
                metadata:
                  type: object
              minProperties: 1
              example:
                is_active: false
      responses:
        '200':
          description: Subscription plan updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionPlan'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Delete subscription plan
      description: Delete a subscription plan belonging to a merchant
      operationId: deleteSubscriptionPlan
      parameters:
        - name: merchant_id
          in: path
          required: true
          description: The ID of the merchant.
          schema:
            type: string
            format: uuid
        - name: plan_id
          in: path
          required: true
          description: The ID of the subscription plan.
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Subscription plan deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /transactions:
    get:
      summary: List transactions for a merchant
      operationId: transactionsList
      parameters:
        - name: merchant_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
  /providers:
    get:
      summary: List available payment providers
      operationId: providers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Provider'
        '401':
          $ref: '#/components/responses/UnauthorizedError' 
  /refunds:
    post:
      summary: Create a refund
      description: Initiate a refund for a transaction
      operationId: createRefund
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRefund'
      responses:
        '201':
          description: Refund created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Refund'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /refunds/{refund_id}:
    get:
      summary: Get refund details
      description: Get details of a specific refund
      operationId: getRefund
      parameters:
        - name: refund_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Refund details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Refund'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      summary: Update refund status
      description: Update the status of a refund
      operationId: updateRefund
      parameters:
        - name: refund_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [completed, failed]
                metadata:
                  type: object
              required:
                - status
      responses:
        '200':
          description: Refund updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Refund'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' 
  /customers:
    post:
      summary: Create a customer
      description: Create a new customer for a merchant
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomer'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    get:
      summary: List customers
      description: List all customers for a merchant
      operationId: listCustomers
      parameters:
        - name: merchant_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: email
          in: query
          schema:
            type: string
            format: email
        - name: phone_number
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /customers/{customer_id}:
    get:
      summary: Get customer details
      description: Get details of a specific customer
      operationId: getCustomer
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Customer details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      summary: Update customer
      description: Update customer details
      operationId: updateCustomer
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                phone_number:
                  type: string
                first_name:
                  type: string
                last_name:
                  type: string
                address:
                  type: object
                  properties:
                    street:
                      type: string
                    city:
                      type: string
                    state:
                      type: string
                    postal_code:
                      type: string
                    country:
                      type: string
                metadata:
                  type: object
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      summary: Delete customer
      description: Delete a customer
      operationId: deleteCustomer
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Customer deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError' 
  /payment-links:
    post:
      summary: Create a payment link
      description: Create a new payment link for collecting payments
      operationId: createPaymentLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentLink'
      responses:
        '201':
          description: Payment link created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentLink'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    get:
      summary: List payment links
      description: List all payment links for a merchant
      operationId: listPaymentLinks
      parameters:
        - name: merchant_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: link_type
          in: query
          schema:
            $ref: '#/components/schemas/PaymentLinkType'
        - name: currency_code
          in: query
          schema:
            $ref: '#/components/schemas/CurrencyCode'
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of payment links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentLink'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /webhooks:
    post:
      summary: Register a new webhook endpoint
      description: Create a new webhook endpoint for receiving event notifications
      operationId: createWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhook'
      responses:
        '201':
          description: Webhook endpoint created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    get:
      summary: List webhook endpoints
      description: List all webhook endpoints for a merchant
      operationId: listWebhooks
      parameters:
        - name: merchant_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: organization_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of webhook endpoints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Webhook'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...